# src/router/yaml/swagger.yaml
openapi: 3.0.0
info:
    title: IDika API
    version: 1.0.0
    description: API
servers:
    - url: http://localhost:3000/api/v1
      description: Local server
    - url: https://idika-api.herokuapp.com/api/v1
      description: Heroku server
tags:
    - name: Anime
      description: API for searching, retrieving details, and getting top anime
host: localhost:3000
paths:
    /anime/search:
        get:
            summary: Mencari anime berdasarkan judul
            tags:
                - Anime
            parameters:
                - name: title
                  in: query
                  description: Judul anime yang akan dicari
                  required: true
            responses:
                '200':
                    description: Data anime yang ditemukan
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/AnimeSearchResponse'
                '400':
                    $ref: '#/components/responses/BadRequest'
    /anime/detail/{id}:
        get:
            summary: Mendapatkan detail anime berdasarkan ID
            tags:
                - Anime
            parameters:
                - name: id
                  in: path
                  description: ID anime yang akan dilihat detailnya
                  required: true
            responses:
                '200':
                    description: Data detail anime
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/AnimeDetailResponse'
                '400':
                    $ref: '#/components/responses/BadRequest'
    /anime/top:
        get:
            summary: Mendapatkan daftar top anime
            tags:
                - Anime
            parameters:
                - name: type
                  in: query
                  description: Tipe daftar top anime (e.g., airing, upcoming)
                  required: true
                - name: page
                  in: query
                  description: Halaman yang akan ditampilkan
                  required: false
                  default: 1
                - name: top
                  in: query
                  description: Jumlah anime per halaman
                  required: false
                  default: 50
            responses:
                '200':
                    description: Data daftar top anime
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/AnimeTopResponse'
                '400':
                    $ref: '#/components/responses/BadRequest'
components:
    responses:
        BadRequest:
            description: Permintaan tidak valid
            content:
                application/json:
                    schema:
                        $ref: '#/components/schemas/BadRequest'
    schemas:
        AnimeSearchResponse:
            type: object
            properties:
                success:
                    type: boolean
                    example: true
                statusCode:
                    type: integer
                    example: 200
                message:
                    type: string
                    example: Request processed successfully
                error:
                    type: null
                    example: null
                data:
                    type: array
                    items:
                        type: object
                        properties:
                            # Example of title: Naruto in myanimelist
                            title:
                                type: string
                                example: Naruto
                            id:
                                type: string
                                example: 20
                            url:
                                type: string
                                example: https://myanimelist.net/anime/20/Naruto
                            detail:
                                type: string
                                example: http://localhost:3000/api/v1/anime/detail/20
        AnimeDetailResponse:
        # Example of anime detail: Naruto in myanimelist
            type: object
            properties:
                success:
                    type: boolean
                    example: true
                statusCode:
                    type: integer
                    example: 200
                message:
                    type: string
                    example: Request processed successfully
                error:
                    type: null
                    example: null
                data:
                    type: object
                    properties:
                        title:
                            type: object
                            properties:
                                original:
                                    type: string
                                    example: Naruto
                                english:
                                    type: string
                                    example: Naruto
                                japanese:
                                    type: string
                                    example: ナルト
                                synonyms:
                                    type: array
                                    items:
                                        type: string
                                    example: ['NARUTO']
                        id:
                            type: string
                            example: 20
                        url:
                            type: string
                            example: https://myanimelist.net/anime/20/Naruto
                        detail:
                            type: object
                            properties:
                                type:
                                    type: string
                                    example: TV
                                episodes:
                                    type: integer
                                    example: 220
                                status:
                                    type: string
                                    example: Finished Airing
                                aired:
                                    type: object
                                    properties:
                                        from:
                                            type: string
                                            example: Oct 3, 2002
                                        to:
                                            type: string
                                            example: Feb 8, 2007
                                premiered:
                                    type: string
                                    example: Fall 2002
                                broadcast:
                                    type: string
                                    example: Thursdays at 19:30 (JST)
                                producers:
                                    type: array
                                    items:
                                        type: string
                                    example: ['TV Tokyo', 'Aniplex', 'Shueisha']
                                licensors:
                                    type: array
                                    items:
                                        type: string
                                    example: ['Viz Media']
                                studios:
                                    type: array
                                    items:
                                        type: string
                                    example: ['Studio Pierrot']
                                source:
                                    type: string
                                    example: Manga
                                genres:
                                    type: array
                                    items:
                                        type: string
                                    example: ['Action', 'Adventure', 'Comedy', 'Super Power', 'Martial Arts', 'Shounen']
                                themes:
                                    type: array
                                    items:
                                        type: string
                                    example: ['Ninja']
                                duration:
                                    type: string
                                    example: '23 min. per ep.'
                                rating:
                                    type: string
                                    example: PG-13 - Teens 13 or older
                                statistic:
                                    type: object
                                    properties:
                                        score:
                                            type: object
                                            properties:
                                                value:
                                                    type: string
                                                    example: "7.91"
                                                scoredBy:
                                                    type: string
                                                    example: "718.409"
                                        ranked:
                                            type: string
                                            example: "Ranked #112"
                                        popularity:
                                            type: string
                                            example: "4"
                                        members:
                                            type: string
                                            example: "1.174.747"
                                        favorites:
                                            type: string
                                            example: "68.711"
                        song:
                            type: object
                            properties:
                                opening:
                                    $ref: '#/components/schemas/OpeningSong'
                                ending:
                                    $ref: '#/components/schemas/EndingSong'
        AnimeTopResponse:
        # Example of top anime: Airing in myanimelist
            type: object
            properties:
                success:
                    type: boolean
                    example: true
                statusCode:
                    type: integer
                    example: 200
                message:
                    type: string
                    example: Request processed successfully
                error:
                    type: null
                    example: null
                data:
                    type: array
                    items:
                        type: object
                        properties:
                            rank:
                                type: integer
                                example: 1
                            title:
                                type: string
                                example: "Shingeki no Kyojin: The Final Season"
                            id:
                                type: string
                                example: "40028"
                            url:
                                type: string
                                example: "https://myanimelist.net/anime/40028/Shingeki_no_Kyojin__The_Final_Season"
                            score:
                                type: string
                                example: "9.11"
        OpeningSong:
        # Example of opening song: Naruto in myanimelist
            type: array
            items:
                type: object
                properties:
                    title:
                        type: string
                        example: "R★O★C★K★S (performed by Hound Dog)"
                    episode:
                        type: string
                        example: "1-25"
                    artist:
                        type: string
                        example: "Hound Dog"
                    play:
                        type: object
                        properties:
                            spotify:
                                type: string
                                example: "https://open.spotify.com/track/6ZCk2vkpQ7QfGfY9YkRZuI?si=6ZCk2vkpQ7QfGfY9YkRZuI"
                            youtubeMusic:
                                type: string
                                example: "https://music.youtube.com/watch?v=6ZCk2vkpQ7Q&feature=share"
                            appleMusic:
                                type: string
                                example: "https://music.apple.com/us/album/r-o-c-k-s/1440838475?i=1440838476"
                            amazonMusic:
                                type: string
                                example: "https://music.amazon.com/albums/B07JH1YQ1Z?trackAsin=B07JH1YQ1Z&ref=dm_sh_6ZCk2vkpQ7QfGfY9YkRZuI"
        EndingSong:
        # Example of ending song: Naruto in myanimelist
            type: array
            items:
                type: object
                properties:
                    title:
                        type: string
                        example: "Wind"
                    episode:
                        type: string
                        example: "1-25"
                    artist:
                        type: string
                        example: "Akeboshi"
                    play:
                        type: object
                        properties:
                            spotify:
                                type: string
                                example: "https://open.spotify.com/track/6ZCk2vkpQ7QfGfY9YkRZuI"
                            youtubeMusic:
                                type: string
                                example: "https://music.youtube.com/watch?v=6ZCk2vkpQ7QfGfY9YkRZuI"
                            appleMusic:
                                type: string
                                example: "https://music.apple.com/us/album/wind/1440837783?i=1440837784"
                            amazonMusic:
                                type: string
                                example: "https://music.amazon.com/albums/B07JN6LZQH?trackAsin=B07JN6LZQH&ref=dm_sh_2e8a-2e8a-dmcp-8e8e-3b1a1&musicTerritory=US&marketplaceId=ATVPDKIKX0DER"
        BadRequest:
            type: object
            properties:
                success:
                    type: boolean
                    example: false
                statusCode:
                    type: integer
                    example: 400
                message:
                    type: string
                    example: Request processed failed
                error:
                    type: object
                    properties:
                        code:
                            type: string
                            example: bad_request
                        message:
                            type: string
                            example: Bad request
                        details:
                            type: array
                            items:
                                type: object
                                properties:
                                    code:
                                        type: string
                                        example: invalid_query_value | missing_query | invalid_body_value | missing_parameter | invalid_parameter_value
                                    message:
                                        type: string
                                        example: Invalid query value | Missing query | Invalid body value | Missing parameter | Invalid parameter value
                data:
                    type: null
                    example: null